SessionTest: split magic method specific tests to dedicated test class
